Some small improvements

+++ task5 - Future Java/src/main/java/com/codecool/spingboot_tasks/future_java/controller/LongCalculationsController.java
@@ -16,7 +16,7 @@ public class LongCalculationsController {
     }
 
     @GetMapping("/calc")
-    public BigInteger calc(){
+    public int calc(){
         return service.calc();
     }
 }
